+++ dynamic_accounts_report/static/src/js/general_ledger.js
@@ -9,6 +9,9 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
     var QWeb = core.qweb;
     var _t = core._t;
 
+    var datepicker = require('web.datepicker');
+    var time = require('web.time');
+
     window.click_num = 0;
     var GeneralLedger = AbstractAction.extend({
     template: 'GeneralTemp',
@@ -20,6 +23,8 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
             'click #xlsx': 'print_xlsx',
             'click .gl-line': 'show_drop_down',
             'click .view-account-move': 'view_acc_move',
+                        'mousedown div.input-group.date[data-target-input="nearest"]': '_onCalendarIconClick',
+
         },
 
         init: function(parent, action) {
@@ -58,6 +63,35 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
             }
         },
 
+        _onCalendarIconClick: function (ev) {
+        var $calendarInputGroup = $(ev.currentTarget);
+
+        var calendarOptions = {
+
+        minDate: moment({ y: 1000 }),
+            maxDate: moment().add(200, 'y'),
+            calendarWeeks: true,
+            defaultDate: moment().format(),
+            sideBySide: true,
+            buttons: {
+                showClear: true,
+                showClose: true,
+                showToday: true,
+            },
+
+            icons : {
+                date: 'fa fa-calendar',
+
+            },
+            locale : moment.locale(),
+            format : time.getLangDateFormat(),
+             widgetParent: 'body',
+             allowInputToggle: true,
+        };
+
+        $calendarInputGroup.datetimepicker(calendarOptions);
+    },
+
 
         load_data: function (initial_render = true) {
             var self = this;
@@ -70,15 +104,15 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
                         method: 'view_report',
                         args: [[this.wizard_id], action_title],
                     }).then(function(datas) {
-                    _.each(datas['report_lines'], function(rep_lines) {
-                            rep_lines.debit = self.format_currency(datas['currency'],rep_lines.debit);
-                            rep_lines.credit = self.format_currency(datas['currency'],rep_lines.credit);
-                            rep_lines.balance = self.format_currency(datas['currency'],rep_lines.balance);
-
-
-
-
-                            });
+//                    _.each(datas['report_lines'], function(rep_lines) {
+//                            rep_lines.debit = self.format_currency(datas['currency'],rep_lines.debit);
+//                            rep_lines.credit = self.format_currency(datas['currency'],rep_lines.credit);
+//                            rep_lines.balance = self.format_currency(datas['currency'],rep_lines.balance);
+//
+//
+//
+//
+//                            });
 
                             if (initial_render) {
                                     self.$('.filter_view_tb').html(QWeb.render('GLFilterView', {
@@ -96,6 +130,9 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
                                     });
                                     self.$el.find('.analytic_tags').select2({
                                         placeholder: 'Analytic Tags...',
+                                    });
+                                    self.$el.find('.target_move').select2({
+                                        placeholder: 'Target Move...',
                                     });
 
                             }
@@ -234,21 +271,21 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
                     var action_title = self._title
                     self._rpc({
                         model: 'account.general.ledger',
-                        method: 'view_report',
+                        method: 'get_accounts_line',
                         args: [
-                            [self.wizard_id], action_title
+                            [self.wizard_id], account_id, action_title
                         ],
                     }).then(function(data) {
-                    _.each(data['report_lines'], function(rep_lines) {
-                            _.each(rep_lines['move_lines'], function(move_line) {
-
-                             move_line.debit = self.format_currency(data['currency'],move_line.debit);
-                            move_line.credit = self.format_currency(data['currency'],move_line.credit);
-                            move_line.balance = self.format_currency(data['currency'],move_line.balance);
-
-
-                             });
-                             });
+//                    _.each(data['report_lines'], function(rep_lines) {
+//                            _.each(rep_lines['move_lines'], function(move_line) {
+//
+//                             move_line.debit = self.format_currency(data['currency'],move_line.debit);
+//                            move_line.credit = self.format_currency(data['currency'],move_line.credit);
+//                            move_line.balance = self.format_currency(data['currency'],move_line.balance);
+//
+//
+//                             });
+//                             });
 
                     for (var i = 0; i < data['report_lines'].length; i++) {
 
@@ -321,7 +358,6 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
 
             var account_ids = [];
             var account_text = [];
-
             var account_res = document.getElementById("acc_res")
             var account_list = $(".account").select2('data')
             for (var i = 0; i < account_list.length; i++) {
@@ -342,8 +378,7 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
             }
             filter_data_selected.account_ids = account_ids
 
-
-
+            if (self._title == "General Ledger"){
              var journal_ids = [];
             var journal_text = [];
             var journal_res = document.getElementById("journal_res")
@@ -367,6 +402,9 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
             }
             filter_data_selected.journal_ids = journal_ids
 
+            }
+
+
             var analytic_ids = []
             var analytic_text = [];
             var analytic_res = document.getElementById("analytic_res")
@@ -389,7 +427,6 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
 
             }
             filter_data_selected.analytic_ids = analytic_ids
-
             var analytic_tag_ids = []
             var analytic_tag_text = [];
             var analytic_tag_res = document.getElementById("analytic_tag_res")
@@ -412,22 +449,31 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
             }
             filter_data_selected.analytic_tag_ids = analytic_tag_ids
 
-            if ($("#date_from").val()) {
+//            if ($("#date_from").val()) {
+//
+//                var dateString = $("#date_from").val();
+//                filter_data_selected.date_from = dateString;
+//            }
+//            if ($("#date_to").val()) {
+//                var dateString = $("#date_to").val();
+//                filter_data_selected.date_to = dateString;
+//            }
 
-                var dateString = $("#date_from").val();
-                filter_data_selected.date_from = dateString;
+ if (this.$el.find('.datetimepicker-input[name="gen_date_from"]').val()) {
+                filter_data_selected.date_from = moment(this.$el.find('.datetimepicker-input[name="gen_date_from"]').val(), time.getLangDateFormat()).locale('en').format('YYYY-MM-DD');
             }
-            if ($("#date_to").val()) {
-                var dateString = $("#date_to").val();
-                filter_data_selected.date_to = dateString;
+
+            if (this.$el.find('.datetimepicker-input[name="gen_date_to"]').val()) {
+                filter_data_selected.date_to = moment(this.$el.find('.datetimepicker-input[name="gen_date_to"]').val(), time.getLangDateFormat()).locale('en').format('YYYY-MM-DD');
             }
+
 
             if ($(".target_move").length) {
             var post_res = document.getElementById("post_res")
-            filter_data_selected.target_move = $(".target_move")[0].value
-            post_res.value = $(".target_move")[0].value
+            filter_data_selected.target_move = $(".target_move")[1].value
+            post_res.value = $(".target_move")[1].value
                     post_res.innerHTML=post_res.value;
-              if ($(".target_move")[0].value == "") {
+              if ($(".target_move")[1].value == "") {
               post_res.innerHTML="posted";
 
               }