+++ dynamic_accounts_report/static/src/js/trial_balance.js
@@ -9,6 +9,9 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
     var QWeb = core.qweb;
     var _t = core._t;
 
+    var datepicker = require('web.datepicker');
+    var time = require('web.time');
+
     window.click_num = 0;
     var TrialBalance = AbstractAction.extend({
     template: 'TrialTemp',
@@ -19,6 +22,8 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
             'click #pdf': 'print_pdf',
             'click #xlsx': 'print_xlsx',
             'click .show-gl': 'show_gl',
+                        'mousedown div.input-group.date[data-target-input="nearest"]': '_onCalendarIconClick',
+
         },
 
         init: function(parent, action) {
@@ -44,6 +49,35 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
             })
         },
 
+        _onCalendarIconClick: function (ev) {
+        var $calendarInputGroup = $(ev.currentTarget);
+
+        var calendarOptions = {
+
+//        minDate: moment({ y: 1000 }),
+//            maxDate: moment().add(200, 'y'),
+//            calendarWeeks: true,
+//            defaultDate: moment().format(),
+//            sideBySide: true,
+//            buttons: {
+//                showClear: true,
+//                showClose: true,
+//                showToday: true,
+//            },
+
+            icons : {
+                date: 'fa fa-calendar',
+
+            },
+            locale : moment.locale(),
+            format : time.getLangDateFormat(),
+             widgetParent: 'body',
+             allowInputToggle: true,
+        };
+
+        $calendarInputGroup.datetimepicker(calendarOptions);
+    },
+
 
         load_data: function (initial_render = true) {
             var self = this;
@@ -55,6 +89,8 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
                         method: 'view_report',
                         args: [[this.wizard_id]],
                     }).then(function(datas) {
+
+
                             _.each(datas['report_lines'], function(rep_lines) {
                             rep_lines.debit = self.format_currency(datas['currency'],rep_lines.debit);
                             rep_lines.credit = self.format_currency(datas['currency'],rep_lines.credit);
@@ -69,6 +105,9 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
                                     }));
                                     self.$el.find('.journals').select2({
                                         placeholder: 'Select Journals...',
+                                    });
+                                    self.$el.find('.target_move').select2({
+                                        placeholder: 'Target Move...',
                                     });
                             }
                             var child=[];
@@ -223,21 +262,29 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
             }
             filter_data_selected.journal_ids = journal_ids
 
-            if ($("#date_from").val()) {
-                var dateString = $("#date_from").val();
-                filter_data_selected.date_from = dateString;
+//            if ($("#date_from").val()) {
+//                var dateString = $("#date_from").val();
+//                filter_data_selected.date_from = dateString;
+//            }
+//            if ($("#date_to").val()) {
+//                var dateString = $("#date_to").val();
+//                filter_data_selected.date_to = dateString;
+//            }
+
+if (this.$el.find('.datetimepicker-input[name="date_from"]').val()) {
+                filter_data_selected.date_from = moment(this.$el.find('.datetimepicker-input[name="date_from"]').val(), time.getLangDateFormat()).locale('en').format('YYYY-MM-DD');
             }
-            if ($("#date_to").val()) {
-                var dateString = $("#date_to").val();
-                filter_data_selected.date_to = dateString;
+
+            if (this.$el.find('.datetimepicker-input[name="date_to"]').val()) {
+                filter_data_selected.date_to = moment(this.$el.find('.datetimepicker-input[name="date_to"]').val(), time.getLangDateFormat()).locale('en').format('YYYY-MM-DD');
             }
 
             if ($(".target_move").length) {
             var post_res = document.getElementById("post_res")
-            filter_data_selected.target_move = $(".target_move")[0].value
-            post_res.value = $(".target_move")[0].value
+            filter_data_selected.target_move = $(".target_move")[1].value
+            post_res.value = $(".target_move")[1].value
                     post_res.innerHTML=post_res.value;
-              if ($(".target_move")[0].value == "") {
+              if ($(".target_move")[1].value == "") {
               post_res.innerHTML="posted";
 
               }