+++ dynamic_accounts_report/doc/RELEASE_NOTES.md
@@ -21,5 +21,28 @@
 #### UPDT
 - Updated
 
+#### 10.08.2021
+#### Version 14.0.1.2.3
+#### UPDT
+- Style Issue Updated
 
 
+#### 20.12.2021
+#### Version 14.0.1.2.4
+#### UPDT
+- Translation issue and Calendar format issue
+
+#### 15.01.2022
+#### Version 14.0.1.2.5
+#### UPDT
+- Arabic Translation added
+
+#### 01.02.2022
+#### Version 14.0.1.2.6
+#### UPDT AND BUGFIX
+- Multi-company and Translation Update and Bugfix
+
+#### 16.04.2022
+#### Version 14.0.1.2.7
+#### UPDT AND BUGFIX
+- Loading Issue and orderby date